+++ CMakeLists.txt
@@ -69,7 +69,7 @@ SET(TARGET_SRC
      include/exadg/time_integration/bdf_constants.cpp
      include/exadg/time_integration/ab_constants.cpp
      include/exadg/time_integration/am_constants.cpp
-     include/exadg/time_integration/extrapolation_scheme.cpp
+     include/exadg/time_integration/extrapolation_constants.cpp
      include/exadg/time_integration/time_int_base.cpp
      include/exadg/time_integration/time_int_bdf_base.cpp
      include/exadg/time_integration/time_int_explicit_runge_kutta_base.cpp
